refactor: make dropdown translatable
