+++ src/utils/color-formats.js
@@ -1,0 +1,32 @@
+export const colorFormats = [
+  {
+    key: "name",
+    // Color name
+    description: _("Name"),
+  },
+  {
+    key: "hex",
+    // Hexadecimal color format
+    description: _("Hex"),
+  },
+  {
+    key: "rgb",
+    // RGB color format
+    description: _("RGB"),
+  },
+  {
+    key: "rgb_percent",
+    // RGB Percent color format
+    description: _("RGB Percent"),
+  },
+  {
+    key: "hsl",
+    // HSL color format
+    description: _("HSL"),
+  },
+  {
+    key: "hsv",
+    // HSV color format
+    description: _("HSV"),
+  },
+];
